:recycle: Refactor: 위치 이동 및 QClass import 수정
<footer>
- 관련: #134
