--- application/external-api/src/main/java/com/namo/spring/application/external/domain/group/dao/repository/group/MoimRepositoryCustom.java
@@ -1,9 +1,0 @@
-package com.example.namo2.domain.group.dao.repository.group;
-
-import com.example.namo2.domain.group.domain.Moim;
-
-public interface MoimRepositoryCustom {
-	Moim findMoimHavingLockWithMoimAndUsersByCode(String code);
-
-	Moim findHavingLockById(Long moimId);
-}
