--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/dao/repository/alarm/AlarmRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.example.namo2.domain.individual.dao.repository.alarm;
-
-public interface AlarmRepositoryCustom {
-}
