--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/dao/repository/image/ImageRepositoryCustom.java
@@ -1,7 +1,0 @@
-package com.example.namo2.domain.individual.dao.repository.image;
-
-import com.example.namo2.domain.individual.domain.Schedule;
-
-public interface ImageRepositoryCustom {
-	void deleteDiaryImages(Schedule schedule);
-}
