--- application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/dao/repository/palette/PaletteRepository.java
@@ -1,8 +1,0 @@
-package com.example.namo2.domain.individual.dao.repository.palette;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.namo2.domain.individual.domain.Palette;
-
-public interface PaletteRepository extends JpaRepository<Palette, Long> {
-}
