+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/repository/image/ImageRepositoryCustom.java
@@ -1,0 +1,7 @@
+package com.namo.spring.application.external.domain.individual.repository.image;
+
+import com.namo.spring.application.external.domain.individual.domain.Schedule;
+
+public interface ImageRepositoryCustom {
+	void deleteDiaryImages(Schedule schedule);
+}
