+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/repository/palette/PaletteRepository.java
@@ -1,0 +1,8 @@
+package com.namo.spring.application.external.domain.individual.repository.palette;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.namo.spring.application.external.domain.individual.domain.Palette;
+
+public interface PaletteRepository extends JpaRepository<Palette, Long> {
+}
