--- application/external-api/src/main/java/com/namo/spring/application/external/domain/user/dao/repository/TermRepository.java
@@ -1,16 +1,0 @@
-package com.example.namo2.domain.user.dao.repository;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.example.namo2.domain.user.domain.constant.Content;
-import com.example.namo2.domain.user.domain.Term;
-import com.example.namo2.domain.user.domain.User;
-
-public interface TermRepository extends JpaRepository<Term, Long> {
-	Optional<Term> findTermByContentAndUser(Content content, User user);
-
-	List<Term> findTermsByUser(User user);
-}
